+++ source/index.js
@@ -3,6 +3,37 @@ import { render }        from 'react-dom';
 
 import MiniMongoExplorer from './components/MiniMongoExplorer';
 
-window.MiniMongoExplorer = minimongo =>
-    render(createElement(MiniMongoExplorer, { minimongo }), document.body)
-;
+document.addEventListener('DOMContentLoaded', () => {
+    const ISODate = /^(\d{4})-(\d{2})-(\d{2})T(\d{2}):(\d{2}):(\d{2}(?:\.\d*))(?:Z|(\+|-)([\d|:]*))?$/;
+
+    render(createElement(MiniMongoExplorer, {
+        fetch: callback =>
+            chrome.devtools.inspectedWindow.eval(`
+                (function () {
+                    if (typeof Meteor === 'undefined') {
+                        return '{}';
+                    }
+
+                    let collections = Meteor.connection._mongo_livedata_collections;
+                    if (collections) {
+                        collections = Object.keys(collections)
+                            .reduce((minimongo, collection) =>
+                                ((minimongo[collection] = collections[collection]._docs._map), minimongo)
+                            , {});
+                    } else {
+                        collections = {};
+                    }
+
+                    return JSON.stringify(collections);
+                })()
+            `, result =>
+                callback(JSON.parse(result, (key, value) =>
+                    typeof value === 'string'
+                        ? ISODate.test(value)
+                            ? new Date(value)
+                            : value
+                        : value
+                ))
+            )
+    }), document.body);
+});