+++ webpack.config.babel.js
@@ -18,7 +18,6 @@ export default {
     plugins: [
         new webpack.DefinePlugin({ 'process.env.NODE_ENV': '"production"' }),
         new webpack.LoaderOptionsPlugin({ minimize: true }),
-        new webpack.optimize.DedupePlugin(),
         new webpack.optimize.UglifyJsPlugin({ compress: { warnings: false }, sourceMap: false }),
         new webpack.optimize.OccurrenceOrderPlugin()
     ]